travis: switch to oraclejdk8
There is a weird error when bnd runs, namely:

  In component nu.muntea.custode.sensor.cpu_temp.internal.TemperatureRecorder,
  @Reference cannot be used for method parameters

Which is weird since the bnd code explicitly checks for the method
being a constructor, which it is. Maybe the problem is that Travis
uses a mixed setup, revealed from the build log

  $ java -Xmx32m -version
  openjdk version "1.8.0_141"
  OpenJDK Runtime Environment (build 1.8.0_141-8u141-b15-3~14.04-b15)
  OpenJDK 64-Bit Server VM (build 25.141-b15, mixed mode)

  $ javac -J-Xmx32m -version
  javac 9.0.1

so let's try and force a different version.